+++ CHANGELOG.md
@@ -13,6 +13,14 @@
 
 ---
 
+## [v1.1.0](https://github.com/foxifyjs/i18n/releases/tag/v1.1.0) - _(2019-05-29)_
+
+- :zap: Added ability to define optional parameter
+- :zap: Added ability to define default value for parameters
+- :zap: Added ability to reference a translation inside another translation
+- :zap: Added ability to choose one of [`ignore`, `warn`, `error`] severities for `setLocale` and `translate` method
+- :zap: Added ability to set to strict mode both in `config` and when creating new instance
+
 ## [v1.0.1](https://github.com/foxifyjs/i18n/releases/tag/v1.0.1) - _(2019-05-24)_
 
 - :beetle: Fixed the typescript declaration bug